refactor(android): cleanup WaitingRoom
