+++ apps/android/app/src/androidTest/java/at/aau/serg/android/lobby/WaitingRoomScreenTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onAllNodesWithText
 import androidx.compose.ui.test.onNodeWithText
-import at.aau.serg.android.ui.lobby.LobbyUiState
+import at.aau.serg.android.ui.screens.lobby.waiting.components.LobbyUiState
 import at.aau.serg.android.ui.screens.lobby.main.LobbyViewModel
 import at.aau.serg.android.ui.screens.lobby.waiting.WaitingRoomScreen
 import at.aau.serg.android.ui.state.LoadState
