--- apps/android/app/src/main/java/at/aau/serg/android/ui/screens/lobby/waiting/components/WaitingRoomActions.kt
@@ -1,39 +1,0 @@
-package at.aau.serg.android.ui.screens.lobby.waiting.components
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-private val Purple = Color(0xFF9D3CFF)
-
-@Composable
-fun WaitingRoomActions(
-    onStart: () -> Unit,
-    onInvite: () -> Unit
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Button(
-            onClick = onStart,
-            modifier = Modifier.weight(1f),
-            colors = ButtonDefaults.buttonColors(containerColor = Purple, contentColor = Color.White)
-        ) {
-            Text("Start Game")
-        }
-
-        Button(
-            onClick = onInvite,
-            modifier = Modifier.weight(1f),
-            colors = ButtonDefaults.buttonColors(containerColor = Purple, contentColor = Color.White)
-        ) {
-            Text("Invite")
-        }
-    }
-}
